Load custom extensions regardless of bundle source
Previously we were only loading custom extensions when the bundle was
defined externally with --file, --cnab-file or --reference. It was being
skipped when the bundle definition came from the claim. This caused
parameter sources to not be evaluated when running actions like `porter
upgrade INSTALLATION`.

I have tweaked how we load bundles to ensure that extensions are always
processed.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>